% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSoccer2014.R
\docType{data}
\name{BRSoccer2014}
\alias{BRSoccer2014}
\title{Table of the 2014th Brazilian Soccer Championship.}
\format{
A data frame with 20 rows and 9 variables
\describe{
\item{Times}{name of the teams}
\item{P}{number of points scored (3 * V + E)}
\item{J}{number of matches played (V + E + D)}
\item{V}{number of victories}
\item{E}{number of ties}
\item{D}{number of defeats}
\item{SG}{goals difference (G - GC)}
\item{G}{goals scored}
\item{GC}{goals suffered}
}
}
\source{
\url{https://www.tabeladobrasileirao.net/2014/serie-a/}
}
\usage{
BRSoccer2014
}
\description{
This data set gives the position of each brazilian soccer team
in the 2014th championship with the number of points, games played,
victories, ties, defeats, goal difference, goals scored and goals suffered.
}
\keyword{datasets}
