% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentData.R
\encoding{UTF-8}
\name{currentData}
\alias{currentData}
\title{Get current data of the Brazilian soccer championship}
\usage{
currentData()
}
\value{
An object with class \emph{tabelaBR} containing the information of the Brazilian soccer championship from the year desired.
}
\description{
Number of points, victories, defeats, ties and more info about the current Brazilian soccer championship
}
\examples{
currentData()
}
