% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tabelaBR.R
\name{print.tabelaBR}
\alias{print.tabelaBR}
\title{Print for \code{tabelaBR} objects}
\usage{
\method{print}{tabelaBR}(x,...)
}
\arguments{
\item{x}{An object with class \emph{tabelaBR}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print objects with class \emph{tabelaBR}.
}
\examples{
tabela = previousData(2019)
print(tabela)
}
\seealso{
\code{\link{previousData}}, \code{\link{currentData}}
}
