% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentData.R
\encoding{UTF-8}
\name{currentData}
\alias{currentData}
\title{Get current data of the brazilian soccer championship}
\usage{
currentData()
}
\value{
A data frame with the current information of the brazilian soccer championship
}
\description{
Number of points, victories, defeats, ties and more info about the current brazilian soccer championship
}
