% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previousData.R
\encoding{UTF-8}
\name{previousData}
\alias{previousData}
\title{Get data from previous years of the brazilian soccer championship}
\usage{
previousData(year)
}
\arguments{
\item{year}{The year of the championship for which the table is desired}
}
\value{
A data frame with the information of the brazilian soccer championship from the year desired.
}
\description{
Number of points, victories, defeats, ties and more info about the brazilian soccer championship of a year from 2015 to 2019
}
\examples{
previousData(2019)
}
